Edit field logic
