+++ lib/api/games/client/updateGames.ts
@@ -1,0 +1,13 @@
+import axios from 'axios'
+import { GameCreationAttributes } from '../../../../db/models/GameModel'
+import { API_URL } from '../../constants'
+
+export default async function updateGames(
+    game_ids: string[],
+    body: Omit<GameCreationAttributes, 'id'>
+) {
+    return await axios.put(
+        `${API_URL}/games?game_ids=${game_ids.join(',')}`,
+        body
+    )
+}
